feat(impl) : complete method 'getEnrollments'

+++ src/main/java/org/arquiweb/dto/DegreeEnrollmentsDTO.java
@@ -4,5 +4,7 @@ package org.arquiweb.dto;
 
 public record DegreeEnrollmentsDTO(
     String name,
-    int enrollmentsAmount
-){ }
+    long enrollmentsAmount
+){
+
+}
